+++ packages/next-document-with-helmet/src/with-helmet.provider.tsx
@@ -3,15 +3,15 @@ import React                            from 'react'
 import { Head, Html, Main, NextScript } from 'next/document'
 import { Helmet }                       from 'react-helmet'
 
-export const withHelmet = () => TargetComponent =>
+export const withHelmet = () => (TargetComponent) =>
   class WithHelmet extends TargetComponent {
     static async getInitialProps(context) {
       const props = await TargetComponent.getInitialProps(context)
 
       const helmet = Helmet.renderStatic()
       const helmetHead = Object.keys(helmet)
-        .filter(el => el !== 'htmlAttributes' && el !== 'bodyAttributes')
-        .map(el => helmet[el].toComponent())
+        .filter((el) => el !== 'htmlAttributes' && el !== 'bodyAttributes')
+        .map((el) => helmet[el].toComponent())
         .flat()
 
       props.head = [...props.head, ...helmetHead]