+++ packages/next-document-with-manifest/src/with-manifest.provider.tsx
@@ -1,7 +1,7 @@
 import Document from 'next/document'
 import React    from 'react'
 
-export const withManifest = () => TargetComponent =>
+export const withManifest = () => (TargetComponent) =>
   class WithManifest extends TargetComponent {
     static async getInitialProps(context) {
       const props = await super.getInitialProps(context)