+++ packages/next-document-with-icons/CHANGELOG.md
@@ -5,7 +5,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## 0.3.15 (2020-05-12)
 
-
 ### Features
 
-* **common:** init document with icons ([ddf9b03](https://github.com/monstrs-lab/nextjs-modules/commit/ddf9b0338866156fa0157711a4f6756581225e4a))
+- **common:** init document with icons ([ddf9b03](https://github.com/monstrs-lab/nextjs-modules/commit/ddf9b0338866156fa0157711a4f6756581225e4a))